Move data association to the particle stage
