+++ module/localisation/FieldLocalisation/data/config/FieldLocalisation.yaml
@@ -19,11 +19,17 @@ initial_covariance: [0.5, 0.5, 0.5]
 # Amount of noise added to the particles (how much world can move to adjust for drift) at each time update
 process_noise: [1e-2, 1e-2, 1e-2]
 
+# Field intersection measurement noise
+measurement_noise: [1e-2, 1e-2]
+
 # Bool to save a csv file of the generated occupancy grid map
 save_map: false
 
 # Minimum number of field line points for a measurement update
 min_observations: 50
+
+# Minimum distance an associated landmark needs to be for measurement update
+min_association_distance: 0.5
 
 # Delay before the particle filter starts (allows odometry to settle)
 start_time_delay: 5
